requirements.txt
streamlit
google-generativeai
python-dotenv
AIzaSyAHI-EHOmBfIbjPnNmm_KGOJf-TP-GSnPE